WIIS-63: mise en place accès pour demande de livraison

+++ src/Controller/DemandeController.php
@@ -32,7 +32,7 @@ class DemandeController extends AbstractController
     private $statutRepository;
 
     /**
-     * @var LignreArticleRepository
+     * @var LigneArticleRepository
      */
     private $ligneArticleRepository;
 
@@ -105,24 +105,28 @@ class DemandeController extends AbstractController
      */
     public function edit(Request $request): Response
     {
-        if (!$request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
-            $utilisateur = $this->utilisateurRepository->find(intval($data["demandeur"]));
-            $emplacement = $this->emplacementRepository->find(intval($data['destination']));
-            $demande = $this->demandeRepository->find($data['demandeId']);
-            $demande
-                ->setUtilisateur($utilisateur)
-                ->setDestination($emplacement);
-            $em = $this->getDoctrine()->getEntityManager();
-            $em->flush();
+        if ($this->isGranted('ROLE_SUPER_CUSTOMER')) {
+            if (!$request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
+                $utilisateur = $this->utilisateurRepository->find(intval($data["demandeur"]));
+                $emplacement = $this->emplacementRepository->find(intval($data['destination']));
+                $demande = $this->demandeRepository->find($data['demandeId']);
+                $demande
+                    ->setUtilisateur($utilisateur)
+                    ->setDestination($emplacement);
+                $em = $this->getDoctrine()->getEntityManager();
+                $em->flush();
 
-            $json = [
-                'entete' => $this->renderView('demande/enteteDemandeLivraison.html.twig', [
-                    'demande' => $demande,
-                ])
-            ];
-            return new JsonResponse($json);
+                $json = [
+                    'entete' => $this->renderView('demande/enteteDemandeLivraison.html.twig', [
+                        'demande' => $demande,
+                    ])
+                ];
+                return new JsonResponse($json);
+            }
+            throw new NotFoundHttpException("404");
+        } else {
+            //TODO CG message erreur (pas les droits pour cette action)
         }
-        throw new NotFoundHttpException("404");
     }
 
     /**
@@ -130,31 +134,35 @@ class DemandeController extends AbstractController
      */
     public function new(Request $request): Response
     {
-        if (!$request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
-            $em = $this->getDoctrine()->getManager();
-            $utilisateur = $this->utilisateurRepository->find($data["demandeur"]);
-            $date = new \DateTime('now');
-            $statut = $this->statutRepository->findOneByCategorieAndStatut(Demande::CATEGORIE, Demande::STATUT_BROUILLON);
-            $destination = $this->emplacementRepository->find($data["destination"]);
-            $demande = new Demande();
-            $demande
-                ->setStatut($statut)
-                ->setUtilisateur($utilisateur)
-                ->setdate($date)
-                //                ->setDateAttendu(new \DateTime($data['dateAttendu']))
-                ->setDestination($destination)
-                ->setNumero("D-" . $date->format('YmdHis'))
-                ->setCommentaire($data["commentaire"]);
-            $em->persist($demande);
-            $em->flush();
+        if ($this->isGranted('ROLE_SUPER_CUSTOMER')) {
+            if (!$request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
+                $em = $this->getDoctrine()->getManager();
+                $utilisateur = $this->utilisateurRepository->find($data["demandeur"]);
+                $date = new \DateTime('now');
+                $statut = $this->statutRepository->findOneByCategorieAndStatut(Demande::CATEGORIE, Demande::STATUT_BROUILLON);
+                $destination = $this->emplacementRepository->find($data["destination"]);
+                $demande = new Demande();
+                $demande
+                    ->setStatut($statut)
+                    ->setUtilisateur($utilisateur)
+                    ->setdate($date)
+                    //                ->setDateAttendu(new \DateTime($data['dateAttendu']))
+                    ->setDestination($destination)
+                    ->setNumero("D-" . $date->format('YmdHis'))
+                    ->setCommentaire($data["commentaire"]);
+                $em->persist($demande);
+                $em->flush();
 
-            $data = [
-                "redirect" => $this->generateUrl('ligne_article_show', ['id' => $demande->getId()])
-            ];
+                $data = [
+                    "redirect" => $this->generateUrl('ligne_article_show', ['id' => $demande->getId()])
+                ];
 
-            return new JsonResponse($data);
+                return new JsonResponse($data);
+            }
+            throw new NotFoundHttpException("404");
+        } else {
+            //TODO CG message erreur (pas les droits pour cette action)
         }
-        throw new NotFoundHttpException("404");
     }
 
     /**
@@ -174,17 +182,21 @@ class DemandeController extends AbstractController
      */
     public function delete(Request $request): Response
     {
-        if (!$request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
-            $demande = $this->demandeRepository->find($data['demandeId']);
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($demande);
-            $entityManager->flush();
-            $data = [
-                "redirect" => $this->generateUrl('demande_index')
-            ];
-            return new JsonResponse($data);
+        if ($this->isGranted('ROLE_ADMIN_GT')) {
+            if (!$request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
+                $demande = $this->demandeRepository->find($data['demandeId']);
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($demande);
+                $entityManager->flush();
+                $data = [
+                    "redirect" => $this->generateUrl('demande_index')
+                ];
+                return new JsonResponse($data);
+            }
+            throw new NotFoundHttpException("404");
+        } else {
+            //TODO CG message erreur (pas les droits pour cette action)
         }
-        throw new NotFoundHttpException("404");
     }
 
     /**
